+++ src/components/LabelModal/LabelModal.css
@@ -1,0 +1,20 @@
+.label-header form {
+  display: flex;
+}
+
+.label-modal {
+  position: absolute;
+  top: 30%;
+  left: 40%;
+  width: 20rem;
+  background-color: var(--primary-bg);
+}
+
+.create-label-input {
+  background: transparent;
+  border: none;
+  border-bottom: 1px solid var(--secondary-color);
+  outline: none;
+  color: var(--primary-text);
+  width: 100%;
+}
